Move `get_client` to `memex.search.client`
This will make it easier to test without having to add a
`__init___test.py` file, and it really also belongs into that module.
