+++ src/memex/search/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from memex.search.client import Client
+from memex.search.client import get_client
 from memex.search.config import init
 from memex.search.core import Search
 from memex.search.core import FILTERS_KEY
@@ -11,21 +11,6 @@ __all__ = (
     'get_client',
     'init',
 )
-
-
-def get_client(settings):
-    """Return a client for the Elasticsearch index."""
-    host = settings['es.host']
-    index = settings['es.index']
-    kwargs = {}
-    kwargs['max_retries'] = settings.get('es.client.max_retries', 3)
-    kwargs['retry_on_timeout'] = settings.get('es.client.retry_on_timeout', False)
-    kwargs['timeout'] = settings.get('es.client.timeout', 10)
-
-    if 'es.client_poolsize' in settings:
-        kwargs['maxsize'] = settings['es.client_poolsize']
-
-    return Client(host, index, **kwargs)
 
 
 def includeme(config):
