+++ src/memex/search/client.py
@@ -30,3 +30,18 @@ class Client(object):
                                   # to elasticsearch>=5.0.0.
                                   ca_certs=certifi.where(),
                                   **kwargs)
+
+
+def get_client(settings):
+    """Return a client for the Elasticsearch index."""
+    host = settings['es.host']
+    index = settings['es.index']
+    kwargs = {}
+    kwargs['max_retries'] = settings.get('es.client.max_retries', 3)
+    kwargs['retry_on_timeout'] = settings.get('es.client.retry_on_timeout', False)
+    kwargs['timeout'] = settings.get('es.client.timeout', 10)
+
+    if 'es.client_poolsize' in settings:
+        kwargs['maxsize'] = settings['es.client_poolsize']
+
+    return Client(host, index, **kwargs)
